Changed how completions are handled.
The Atlanta machines have a single dir filled with completion scripts.
All other machines I've used have a script that does the iteration.
This change allows multiple scripts in a single dir to be executed.
NB: This may not work in cases where an iterating script is alongside the
individual scripts.

Signed-off-by: David Panariti <David.Panariti@amd.com>