Renamed old ~/.rc/gnupg for to keep info around.
I need to start tagging the commit with mv-from-to to simulate a
rename just used to prevent the need to dig around inside the repo.
The tag will look just like the rename would've.
Smells like a script.
Need to make sure those tags are always kept and pushed around.