Replaced my lookfor-dir-change with Emacs' `dir-tracker'.
`dir-tracker' is basically the same as mine, except made more robustly and
ready for prime time, not just for me.  Still it uses the hack of extracting
the cwd from the prompt.  I was sad to have to do that.  Maybe I should stick
to mine, since it works and will independent of any changes in `dirtrack'.