Changed the order of args to `dp-toggle-va[lr]'.
The order is far more natural.  The variable to toggle then the way to
toggle. Even though we're not always toggling; we can also force nil and non-nil.
If only the var is specified, then the variable is toggled truth-wise.
I have no idea why I did it the other way.  Maybe I stole from a bad source.
Also, of course, changed all of the callers.

ALSO:
Fixed dp-sel2.

The problem was uncovered by 2 characters with read-only properties in the
kill ring.  Which then revealed the use of an obsolete way of changing
read-only-ness.  Yadda, etc.  Then, a change was made to my toggle var
function signature, which should've waited.  I probably would've failed to
pick the changes apart correctly.