Python change during "upgrade" to Lubuntu 20.04: defecation if imp.
To implib. Hacked quickly, but there's an error that doesn't break anything.
Rule of replacing a lib:
To the maximum extent possible, allow use cases of the old lib to be easily,
quickly and safely replace with new lib.
See the change in the process module (e.g. the amount of grief to replace
popen*()) for a perfect counter example.