Minor changes after massive headache.
Renamed some stuff and committed.  Realized that was wrong and wanted to roll
back the `git way' and lost hours.  Tried to use and understand the magit
author's doc and it was very confusingly written.  Really
non/counter-intuitive names.
Sigh.  Just did what I thought was needed by hand and it seems to have worked.
Signed-off-by: Sound Chaser <davep.git@meduseld.net>