Run the ipython in an xterm.
But it don't work.  The command works in a shell, but not with the binding.
Also, a general problem is that I don't yet put the program name in the
window title since zsh came to live with me.  There's a way to do it, I just
need to find it again.
Jon Lord is sooo much better than Ritchie Blackmore.  Just listen to Highway Star.