Hacked and mangled.  See previous version(s).
There are no real changes other than those related to trying to make the
"official" directory tracker, `dirtrack-mode' handle the new zsh prompt.  The
problem is, that while the prompts look the same in the buffer, the comint
filter hook:
(add-hook 'comint-preoutput-filter-functions 'dirtrack nil t)
sees odd garbage that it can't recognize.  I think that it's some kind of
magic binary escape sequence that the shell interprets when it sees it.