Made umr my current project.
Big flaw? in my multi-host config is when machine names no longer uniquely
identify my work.  E.g. I'm doing AMD work on vilya because of moving out of
cube.  So shit which was predicated on my machines there have to be forced
onto vilya unconditionally in one way or another.
EE.gg. forcing the PROJECT here to be umr.