Added to git.  It was a nightmare getting it to work outside the amd domain.
All of it assumed amd.com and it took a fuck-ton of time to figure it out.
Luckily there was something to base all of the errors on.
I.e. a config that showed how to use Host entries, but assumed the domain
name.  Which will continue to be a biatch until it isn't.
Signed-off-by: David Panariti <david.panariti@amd.com>